Adapt purity tests to Java 25 changes (nested class constructors call Objects.requireNonNull)
