+++ OPAL/tac/src/main/resources/reference.conf
@@ -289,9 +289,10 @@ org.opalj {
           {cf = "java/lang/Thread", m = "isAlive", desc = "()Z", p = "SideEffectFree"},
           {cf = "java/lang/Thread", m = "isInterrupted", desc = "(Z)Z", p = "SideEffectFree"},
           {cf = "java/lang/Throwable", m = "<init>", desc = "*", p = "SideEffectFree"},
-          {cf = "java/lang/Throwable", m = "fillInStackTrace", desc = "(I)Ljava/lang/Throwable;", p = "SideEffectFree"},
+          {cf = "java/lang/Throwable", m = "fillInStackTrace", desc = "()Ljava/lang/Throwable;", p = "SideEffectFree"},
           {cf = "java/lang/Throwable", m = "getStackTraceDepth", desc = "()I", p = "SideEffectFree"},
           {cf = "java/lang/Throwable", m = "getStackTraceElement", desc = "(I)Ljava/lang/StackTraceElement;", p = "SideEffectFree"},
+          {cf = "java/lang/NullPointerException", m = "fillInStackTrace", desc = "()Ljava/lang/Throwable;", p = "SideEffectFree"},
           {cf = "java/util/TimeZone", m = "getSystemGMTOffsetID", desc = "()Ljava/lang/String;", p = "SideEffectFree"},
           {cf = "java/util/TimeZone", m = "getSystemTimeZoneID", desc = "(Ljava/lang/String;)Ljava/lang/String;", p = "SideEffectFree"},
           {cf = "java/util/logging/FileHandler", m = "isSetUID", desc = "()Z", p = "Pure"},
